+++ frontend/client/src/index.js
@@ -1,10 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import DataTable from './ViewComponents/DataGrid';
-import Dashboard from './ViewComponents/Dashboard.js';
 import App from './App.js';
-import DataGridController from './DataGridController';
-import Stage from './DataGridController';
+import Stage from './StaffRecordsDataGridController';
 
 ReactDOM.render(<Stage />, document.getElementById('root'));
